requirements.txt
flask
openai
pinecone-client
langchain
langchain-pinecone
langchain-openai
python-dotenv
gunicorn
